refactor(io): move read/write into dir
