+++ compio-io/src/read/mod.rs
@@ -1,7 +1,8 @@
-use std::io::Result as IoResult;
+use compio_buf::{BufResult, IoBufMut, IoVectoredBufMut, SetBufInit};
 
-use compio_buf::{BufResult, IoBufMut, IoVectoredBufMut};
+mod buf;
 
+pub use buf::*;
 /// AsyncRead
 ///
 /// Async read with a ownership of a buffer
@@ -9,11 +10,13 @@ pub trait AsyncRead {
     /// Read some bytes from this source into the buffer, which implements
     /// [`IoBufMut`], and return a [`BufResult`], consisting of the buffer and a
     /// [`usize`] indicating how many bytes were read.
-    async fn read<T: IoBufMut>(&mut self, buf: T) -> BufResult<usize, T>;
+    async fn read<B: IoBufMut>(&mut self, buf: B) -> BufResult<usize, B>;
 
     /// Like `read`, except that it reads into a type implements
     /// [`IoVectoredBufMut`].
-    async fn read_vectored<T: IoVectoredBufMut>(&mut self, buf: T) -> BufResult<usize, T>;
+    async fn read_vectored<V: IoVectoredBufMut>(&mut self, buf: V) -> BufResult<usize, V>
+    where
+        V::Item: IoBufMut + SetBufInit;
 }
 
 impl<A: AsyncRead + ?Sized> AsyncRead for &mut A {
@@ -21,10 +24,14 @@ impl<A: AsyncRead + ?Sized> AsyncRead for &mut A {
         (**self).read(buf).await
     }
 
-    async fn read_vectored<T: IoVectoredBufMut>(&mut self, buf: T) -> BufResult<usize, T> {
+    async fn read_vectored<T: IoVectoredBufMut>(&mut self, buf: T) -> BufResult<usize, T>
+    where
+        T::Item: IoBufMut + SetBufInit,
+    {
         (**self).read_vectored(buf).await
     }
 }
+
 /// # AsyncReadAt
 ///
 /// Async read with a ownership of a buffer and a position
@@ -36,26 +43,5 @@ pub trait AsyncReadAt {
 impl<A: AsyncReadAt + ?Sized> AsyncReadAt for &mut A {
     async fn read_at<T: IoBufMut>(&mut self, buf: T, pos: usize) -> BufResult<usize, T> {
         (**self).read_at(buf, pos).await
-    }
-}
-
-/// # AsyncBufRead
-///
-/// Async read with buffered content
-pub trait AsyncBufRead: AsyncRead {
-    /// Try fill the internal buffer with data
-    async fn fill_buf(&mut self) -> IoResult<&'_ [u8]>;
-
-    /// Mark how much data is read
-    fn consume(&mut self, amt: usize);
-}
-
-impl<A: AsyncBufRead + ?Sized> AsyncBufRead for &mut A {
-    async fn fill_buf(&mut self) -> IoResult<&'_ [u8]> {
-        (**self).fill_buf().await
-    }
-
-    fn consume(&mut self, amt: usize) {
-        (**self).consume(amt)
     }
 }
